% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geeNET.R
\name{geeNET}
\alias{geeNET}
\title{geeNET}
\usage{
geeNET(form, link, adj.matrix, dataset, SE = "sandwich", wt = NULL,
  toep = 1)
}
\arguments{
\item{form}{the formula for the model to be fit, like in the lm function}

\item{link}{link between xbeta and the mean, identity for normal, logit for logistic, Poissonlog for Poisson with log link,  Gammalog for gamma with log link}

\item{adj.matrix}{adjacency matrix of the network}

\item{dataset}{name of the dataset used for the model}

\item{SE}{choose either the sandwich or model.based SE estimator}

\item{wt}{is the vector of inverse probability weights, defaults to NULL which will give all observations equal weights}

\item{toep}{number of rhos to estimate 1-4 currently are the options for this}
}
\value{
standard regression table output
}
\description{
Fits GEE for data with Network Correlation structures
}
