% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getD.R
\name{getD}
\alias{getD}
\title{getD}
\usage{
getD(betas, x.matrix, link)
}
\arguments{
\item{betas}{p vector of regression parameters}

\item{x.matrix}{nxp matrix of covariates}

\item{link}{link between xbeta and the mean, identify for normal, logit for logistic, Gammalog for gamma (should work)}
}
\value{
matrix of partial derivatives dmu/dbeta
}
\description{
Takes betas and x.matrix and returns matrix of partial derivatives dmu/dbeta
}
