% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMean.R
\name{getMean}
\alias{getMean}
\title{getMean}
\usage{
getMean(betas, x.matrix, link)
}
\arguments{
\item{betas}{p vector of regression parameters}

\item{x.matrix}{nxp matrix of covariates}

\item{link}{link between xbeta and the mean, identify for normal, logit for logistic, Gammalog for gamma with log link}
}
\value{
estimated Mean
}
\description{
Takes betas, x.matrix, and link and returns the estimated mean
}
