% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getV.R
\name{getV}
\alias{getV}
\title{getV}
\usage{
getV(betas, rho, x.matrix, y, adj.matrix, missing.vector, link, W, toep)
}
\arguments{
\item{betas}{p vector of regression parameters}

\item{rho}{the vector of estimated correlations between those connected in the network}

\item{x.matrix}{nxp matrix of covariates}

\item{y}{the vector of dependent variable}

\item{adj.matrix}{adjacency matrix of the network}

\item{missing.vector}{TRUE FALSE vector of missing values}

\item{link}{link between xbeta and the mean, identify for normal, logit for logistic, Poissonlog for Poisson with log link, Gammalog for gamma with log link}

\item{W}{vector of inverse probability weights}

\item{toep}{number of rhos to estimate 1-4 currently are the options for this}
}
\value{
estimated Variance Covariance matrix
}
\description{
Takes betas, correlation vector rho, x.matrix, y, adj.matrix, link, W, toep and returns the variance covariance matrix
}
