% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi.score.R
\name{quasi.score}
\alias{quasi.score}
\title{quasi.score}
\usage{
quasi.score(to.optim, adj.matrix, x.matrix, missing.vector, link, y, W, toep)
}
\arguments{
\item{to.optim}{a p+1 vector with initial values for betas and rho}

\item{adj.matrix}{adjacency matrix of the network}

\item{x.matrix}{nxp matrix of covariates}

\item{missing.vector}{TRUE FALSE vector of missing values}

\item{link}{link between xbeta and the mean, identify for normal, logit for logistic, Poissonlog for Poisson with log link, Gammalog for gamma with log link}

\item{y}{the vector of dependent variable}

\item{W}{vector of inverse probability weights}

\item{toep}{number of rhos to estimate 1-4 currently are the options for this}
}
\value{
betas and rho as a vector
}
\description{
Fits GEE and returns betas and rho vector
}
